+++ NEWS.md
@@ -54,6 +54,9 @@
   (with `false` default) that specifies if columns should be inserted after
   or before `col`.
   ([#2829](https://github.com/JuliaData/DataFrames.jl/pull/2829))
+* `leftjoin!` performing a left join of two data frame objects by updating the
+  left data frame with the joined columns from right data frame.
+  ([#2843](https://github.com/JuliaData/DataFrames.jl/pull/2843))
 * the `DataFrame` constructor when column names are passed to it as a second
   argument now determines if a passed vector of column names is valid based on
   its contents and not element type